+++ src/main/java/com/ayu/realty/member/repository/MemberRepository.java
@@ -1,11 +1,15 @@
-package com.ayu.realty.member;
+package com.ayu.realty.member.repository;
 
-import com.ayu.realty.member.model.Member;
+import com.ayu.realty.member.model.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
 
 @Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
 
     boolean existsByEmail(String email);
+
+    Optional<Member> findByEmail(String email);
 }
